Add flexible action implementation
Tasks can now be composed using any number of actions in arbitrary
order. As part of this change, a new version of the JSON configuration
format is introduced. Tasks stored in the database are upgraded by an
upgrader, while exported task files can be imported using either
format.

This also adds a number of unit tests covering the basic functionality
of all actions.