Fix scheduler being thrown off by manual executions
This fixes an issue where manual execution of tasks outside of the
schedule frequency would prevent tasks from running at the date
where one would expect it to run intuitively.

Previously, users had no control over specific subparameters of the
frequency they selected, so generally speaking, a tasks set to run
annually would always be executed on January 1st around midnight,
assuming that's when the scheduled job runs. With the extended
scheduler, it's now possible to schedule annual tasks to run e.g. in
June. In that scenario, if someone were to manually run the task at
some point before June, the scheduler would treat the task as already
having been executed, so it would not run in June. This is
counterintuitive, so we now run the task at any point where it's due
and hasn't been executed on or after the due date.